Add subtitles and metadata while transcoding the video

+++ lib/apple_tv_converter/media_converter_adapter.rb
@@ -12,20 +12,73 @@ module AppleTvConverter
 
     def transcode(media)
       if convert?(media)
-        puts "* Encoding"
+        puts "* Encoding#{', adding subtitles and metadata' unless media.is_tv_show_episode?}"
 
-        options = get_transcode_options(media)
+        options = {}
+        options[:codecs] = get_transcode_options(media)
+        options[:files] = ""
+        options[:metadata] = ""
+        options[:map] = ""
+        options[:extra] = ""
 
         # Better video and audio transcoding quality
-        if convert_video?(media) || convert_audio?(media)
-          options[:custom] = ''
-          options[:custom] << ' -mbd rd -flags +mv4+aic -trellis 2 -cmp 2 -subcmp 2 -g 300 -pass 1 -qscale 1' if convert_video?(media)
-          options[:custom] << ' -ac 6 -ar 48000 -ab 448k -vol 512' if convert_audio?(media)
+        if convert_video?(media)
+          options[:extra] << ' -mbd rd -flags +mv4+aic -trellis 2 -cmp 2 -subcmp 2 -g 300 -pass 1 -q:v 1 -r 23.98'
         end
 
-        transcoded = media.ffmpeg_data.transcode(media.converted_filename, options) do |progress|
-          printf "\r" + (" " * 40)
-          printf %Q[\r  * Progress: #{(progress * 100).round(2)}%%]
+        if convert_audio?(media)
+          options[:extra] << " -vol 512" # Increase the volume when transcoding
+          options[:extra] << " -ac #{media.ffmpeg_data.audio_channels} -ar #{media.ffmpeg_data.audio_sample_rate} -ab 448k" if media.ffmpeg_data.audio_codec =~ /mp3/i
+          # options << " -q:a 1" if media.ffmpeg_data.audio_codec =~ /ac3/i
+        end
+
+        # If the file is a MKV file, map all tracks when transcoding
+        if media.is_mkv?
+          media.mkv_data.tracks.each do |track|
+            options[:map] << " -map 0:#{track.mkv_info_id}"
+            options[:metadata] << " -metadata:s:#{track.mkv_info_id} language=#{track.language}" if ['audio', 'subtitle', 'subtitles'].include?(track.type)
+          end
+
+          last_stream_id = media.mkv_data.tracks.length - 1
+        else
+          options[:map] << " -map 0:0 -map 0:1"
+          last_stream_id = 1
+        end
+
+        # Load any external subtitle files
+        last_file_id = 0
+        Dir["#{media.original_filename.gsub(/.{4}$/, '.*srt')}"].each do |subtitle_file|
+          last_file_id += 1
+          last_stream_id += 1
+
+          subtitle_language = (subtitle_file.match(/\.(.{3})\.srt$/) || ['unk'])[1]
+          options[:files] << " -i #{Shellwords.escape(subtitle_file)}"
+          options[:metadata] << " -metadata:s:#{last_stream_id} language=#{subtitle_language}"
+          options[:map] << " -map #{last_file_id}:0"
+        end
+
+        # Set metadata tags
+        if media.is_tv_show_episode?
+          options[:metadata] << %Q[ -metadata title="#{media.show} S#{media.season.to_s.rjust(2, '0')}E#{media.number.to_s.rjust(2, '0')}"]
+          options[:metadata] << %Q[ -metadata genre="#{media.genre}"]
+          options[:metadata] << %Q[ -metadata show="#{media.show}"]
+        else
+          options[:metadata] << %Q[ -metadata title="#{media.show}"]
+          options[:metadata] << %Q[ -metadata genre="#{media.quality} Movies"]
+        end
+
+        options = "#{options[:files]} #{options[:codecs]} #{options[:map]} #{options[:metadata]} #{options[:extra]}"
+
+        transcoded = nil
+
+        begin
+          transcoded = media.ffmpeg_data.transcode(media.converted_filename, options) do |progress|
+            printf "\r" + (" " * 40)
+            printf %Q[\r  * Progress: #{(progress * 100).round(2)}%%]
+          end
+        rescue Interrupt
+          puts "\nProcess canceled!"
+          exit!
         end
 
         status = transcoded.valid?
@@ -36,6 +89,7 @@ module AppleTvConverter
           puts "\r  * Progress: [DONE]#{' ' * 20}"
         else
           puts "\r  * Progress: [ERROR]#{' ' * 20}"
+          exit!
         end
       else
         puts "* Encoding: [UNNECESSARY]"
@@ -60,9 +114,17 @@ module AppleTvConverter
     def clean_up(media)
       printf "* Cleaning up"
       begin
-        FileUtils.rm(media.original_filename) unless media.original_filename == media.converted_filename
-        FileUtils.rm(media.artwork_filename) if File.exists?(media.artwork_filename)
-        FileUtils.rm_r list_files(media.original_filename.gsub(File.extname(media.original_filename), '*.srt'))
+        if convert?(media)
+          FileUtils.rm media.original_filename
+
+          if media.converted_filename_equals_original_filename?
+            FileUtils.mv media.converted_filename, media.original_filename
+          end
+
+          FileUtils.rm(media.artwork_filename) if File.exists?(media.artwork_filename)
+          FileUtils.rm_r list_files(media.original_filename.gsub(File.extname(media.original_filename), '*.srt'))
+        end
+
         puts " [DONE]"
       rescue
         puts " [ERROR]"
@@ -76,15 +138,19 @@ module AppleTvConverter
       end
 
       def convert_audio?(media)
-        return media.ffmpeg_data.audio_codec !~ /(?:ac3|aac)/i
+        return media.ffmpeg_data.audio_codec !~ /(?:aac)/i
       end
 
       def convert_video?(media)
         return media.ffmpeg_data.video_codec !~ /.*(?:h264|mpeg4).*/i || media.ffmpeg_data.video_codec =~ /.*(?:xvid|divx).*/i
       end
 
+      def convert_subtitles?(media)
+        (media.is_mkv? && media.mkv_data.tracks.select {|t| t.type == 'subtitle' }.any?) || Dir["#{media.original_filename.gsub(/.{4}$/, '*srt')}"].any?
+      end
+
       def convert?(media)
-        !(media.is_valid? && media.is_mp4?)# && !convert_video?(media) && !convert_audio?(media))
+        !(media.is_valid? && media.is_mp4? && !convert_video?(media) && !convert_audio?(media) && !convert_subtitles?(media))
       end
 
       def has_subtitles?(media)
@@ -92,7 +158,10 @@ module AppleTvConverter
       end
 
       def get_transcode_options(media)
-        raise NotImplementedYetException
+        options = " -vcodec #{convert_video?(media) ? 'mpeg4' : 'copy'} -acodec #{convert_audio?(media) ? 'libfaac' : 'copy'}"
+        options << " -scodec mov_text" if convert_subtitles?(media)
+
+        options
       end
 
       def load_movie_from_imdb(media)
